+++ src/newt/db/_adapter.py
@@ -1,13 +1,8 @@
-import json
-import logging
-import re
 import relstorage.adapters.postgresql
 import relstorage.adapters.postgresql.mover
 import relstorage.adapters.postgresql.schema
 
-from .jsonpickle import JsonUnpickler
-
-logger = logging.getLogger(__name__)
+from .jsonpickle import Jsonifier
 
 class Adapter(relstorage.adapters.postgresql.PostgreSQLAdapter):
 
@@ -30,48 +25,7 @@ class Adapter(relstorage.adapters.postgresql.PostgreSQLAdapter):
         )
         self.connmanager.set_on_store_opened(self.mover.on_store_opened)
 
-skip_class = re.compile('BTrees[.]|ZODB.blob').match
-unicode_surrogates = re.compile(r'\\ud[89a-f][0-9a-f]{2,2}', flags=re.I)
-NoneNoneNone = None, None, None
-def jsonify(oid, data):
-    if not data:
-        return NoneNoneNone
-    unpickler = JsonUnpickler(data)
-    try:
-        klass = json.loads(unpickler.load())
-
-        if isinstance(klass, list):
-            klass, args = klass
-            if isinstance(klass, list):
-                class_name = '.'.join(klass)
-            else:
-                class_name = klass['name']
-        else:
-            class_name = klass['name']
-
-        if skip_class(class_name):
-            return NoneNoneNone
-
-        ghost_pickle = data[:unpickler.pos]
-        state = unpickler.load()
-
-        # xstate = xform(zoid, class_name, state)
-        # if xstate is not state:
-        #     state = xstate
-        #     if not isinstance(state, bytes):
-        #         state = json.dumps(state)
-
-        # Remove unicode surrogate strings, as postgres utf-8
-        # will reject them.
-
-        state = unicode_surrogates.sub(' ', state).replace('\\u0000', ' ')
-    except Exception:
-        logger.warn("Failed pickle load, oid: %r, pickle starts: %r",
-                    oid, data[:50], exc_info=True)
-        return NoneNoneNone
-
-    return class_name, ghost_pickle, state
-
+        self.mover.jsonifier = Jsonifier()
 
 class Mover(relstorage.adapters.postgresql.mover.PostgreSQLObjectMover):
 
@@ -87,7 +41,7 @@ class Mover(relstorage.adapters.postgresql.mover.PostgreSQLObjectMover):
 
     def store_temp(self, cursor, batcher, oid, prev_tid, data):
         super(Mover, self).store_temp(cursor, batcher, oid, prev_tid, data)
-        class_name, ghost_pickle, state = jsonify(oid, data)
+        class_name, ghost_pickle, state = self.jsonifier(oid, data)
         if class_name is None:
             return
         batcher.delete_from('temp_store_json', zoid=oid)
@@ -114,7 +68,7 @@ class Mover(relstorage.adapters.postgresql.mover.PostgreSQLObjectMover):
 
     def restore(self, cursor, batcher, oid, tid, data):
         super(Mover, self).restore(cursor, batcher, oid, tid, data)
-        class_name, ghost_pickle, state = jsonify(oid, data)
+        class_name, ghost_pickle, state = self.jsonifier(oid, data)
         if class_name is None:
             return
         batcher.delete_from('newt', zoid=oid)