+++ CHANGES.rst
@@ -1,18 +1,22 @@
 Changes
 =======
 
-0.2.2 (unreleased)
+0.3.0 (unreleased)
 ==================
 
-- Nothing changed yet.
+- Added an API for following database changes.
 
+- Exposed JSON conversion as an API that can be used to for other
+  applications than updating the newt table, like updating external
+  indexes.
+
+- Created a JSON conversion object to support conversion customization.
 
 0.2.1 (2017-02-06)
 ==================
 
 - Fixed: Packing wasn't handled correctly. Objects removed during
   packing weren't removed from the ``newt`` table.
-
 
 0.2.0 (2017-01-30)
 ==================